+++ src/v2/conversions.rs
@@ -654,11 +654,23 @@ impl TryFrom<ReleaseSchedule> for super::types::AccountReleaseSchedule {
     }
 }
 
-impl TryFrom<AccountVerifyKey> for crate::id::types::VerifyKey {
+impl TryFrom<AccountVerifyKey> for Upward<crate::id::types::VerifyKey> {
     type Error = tonic::Status;
 
-    fn try_from(value: AccountVerifyKey) -> Result<Self, Self::Error> {
-        match value.key.require()? {
+    fn try_from(message: AccountVerifyKey) -> Result<Self, Self::Error> {
+        let key = message
+            .key
+            .map(super::id::types::VerifyKey::try_from)
+            .transpose()?;
+        Ok(Upward::from(key))
+    }
+}
+
+impl TryFrom<account_verify_key::Key> for crate::id::types::VerifyKey {
+    type Error = tonic::Status;
+
+    fn try_from(account_verify_key: account_verify_key::Key) -> Result<Self, Self::Error> {
+        match account_verify_key {
             account_verify_key::Key::Ed25519Key(v) => Ok(Self::Ed25519VerifyKey(consume(&v)?)),
         }
     }
@@ -719,26 +731,30 @@ impl TryFrom<ArThreshold> for crate::id::secret_sharing::Threshold {
     }
 }
 
-impl TryFrom<CredentialPublicKeys> for crate::id::types::CredentialPublicKeys {
+impl TryFrom<CredentialPublicKeys> for Upward<crate::id::types::CredentialPublicKeys> {
     type Error = tonic::Status;
 
     fn try_from(value: CredentialPublicKeys) -> Result<Self, Self::Error> {
-        Ok(Self {
-            keys: value
-                .keys
-                .into_iter()
-                .map(|(k, v)| {
-                    if let Ok(k) = u8::try_from(k) {
-                        let k = k.into();
-                        let v = v.try_into()?;
-                        Ok((k, v))
-                    } else {
-                        Err(tonic::Status::internal("Unexpected key index."))
-                    }
-                })
-                .collect::<Result<_, tonic::Status>>()?,
+        let mut keys = BTreeMap::new();
+
+        for (k, v) in value.keys {
+            let k = u8::try_from(k)
+                .map_err(|_| tonic::Status::internal("Unexpected key index."))?
+                .into();
+
+            let v = match v.try_into()? {
+                Upward::Known(inner) => inner,
+                Upward::Unknown => return Ok(Upward::Unknown), // return immediately if any key is unknown
+            };
+
+            keys.insert(k, v);
+        }
+
+        let creds = crate::id::types::CredentialPublicKeys {
+            keys,
             threshold: value.threshold.require()?.try_into()?,
-        })
+        };
+        Ok(Upward::Known(creds))
     }
 }
 
@@ -856,33 +872,33 @@ impl TryFrom<AccountCredential> for Upward<AccountCredentialWithoutProofs<ArCurv
     type Error = tonic::Status;
 
     fn try_from(message: AccountCredential) -> Result<Self, Self::Error> {
-        let key = message
-            .credential_values
-            .map(super::id::types::AccountCredentialWithoutProofs::try_from)
-            .transpose()?;
-        Ok(Upward::from(key))
-    }
-}
+        let Some(cred) = message.credential_values else {
+            return Ok(Upward::Unknown);
+        };
 
-impl TryFrom<account_credential::CredentialValues>
-    for AccountCredentialWithoutProofs<ArCurve, AttributeKind>
-{
-    type Error = tonic::Status;
-
-    fn try_from(cred: account_credential::CredentialValues) -> Result<Self, Self::Error> {
         match cred {
             account_credential::CredentialValues::Initial(ic) => {
+                let Upward::Known(cred) = ic.keys.require()?.try_into()? else {
+                    return Ok(Upward::Unknown);
+                };
+
                 let icdv = InitialCredentialDeploymentValues {
-                    cred_account: ic.keys.require()?.try_into()?,
+                    cred_account: cred,
                     reg_id: ic.cred_id.require()?.try_into()?,
                     ip_identity: ic.ip_id.require()?.into(),
                     policy: ic.policy.require()?.try_into()?,
                 };
-                Ok(AccountCredentialWithoutProofs::Initial { icdv })
+                Ok(Upward::Known(AccountCredentialWithoutProofs::Initial {
+                    icdv,
+                }))
             }
             account_credential::CredentialValues::Normal(nc) => {
+                let Upward::Known(cred) = nc.keys.require()?.try_into()? else {
+                    return Ok(Upward::Unknown);
+                };
+
                 let cdv = CredentialDeploymentValues {
-                    cred_key_info: nc.keys.require()?.try_into()?,
+                    cred_key_info: cred,
                     cred_id: nc.cred_id.require()?.try_into()?,
                     ip_identity: nc.ip_id.require()?.into(),
                     threshold: nc.ar_threshold.require()?.try_into()?,
@@ -900,7 +916,10 @@ impl TryFrom<account_credential::CredentialValues>
                     policy: nc.policy.require()?.try_into()?,
                 };
                 let commitments = nc.commitments.require()?.try_into()?;
-                Ok(AccountCredentialWithoutProofs::Normal { cdv, commitments })
+                Ok(Upward::Known(AccountCredentialWithoutProofs::Normal {
+                    cdv,
+                    commitments,
+                }))
             }
         }
     }
@@ -1055,6 +1074,7 @@ impl TryFrom<AccountInfo> for super::types::AccountInfo {
             .into_iter()
             .map(|token| token.try_into())
             .collect::<Result<_, _>>()?;
+
         Ok(Self {
             account_nonce,
             account_amount,