+++ CHANGELOG.md
@@ -78,7 +78,8 @@
   - Type `Cooldown` field `status` is now wrapped in `Upward`.
   - Type `BakerEvent::BakerSetOpenStatus` field `open_status` is now wrapped in `Upward`.
   - Type `Event` field `open_status` is now wrapped in `Upward`.
-  - Type `AccountInfo` field `account_credentials` change from `BTreeMap<CredentialIndex,Versioned<AccountCredentialWithoutProofs<ArCurve, AttributeKind>>>` to `BTreeMap<CredentialIndex,Versioned<Upward<AccountCredentialWithoutProofs<ArCurve, AttributeKind>>>>`.
+  - Type `AccountInfo` field `account_credentials` change from `BTreeMap<CredentialIndex,Versioned<AccountCredentialWithoutProofs<ArCurve, AttributeKind>>>` to `BTreeMap<CredentialIndex,Versioned<Upward<AccountCredentialWithoutProofs<ArCurve, AttributeKind>>>>`. The `AccountCredentialWithoutProofs` type is `Unkown` 
+  if an unkown `AccountCredential` type is observed or any of its `verifyKey/publicKey` type is unkown.
 
 ## 7.0.0